refactor(maps): :truck: moved map-comonents files into a new folder components/maps
